Added data structure to replace o(n) lookups
I added a stack structure for the commands to replace the slow O(n)
lookup. It appears to make it around 100x faster. To give context, it
completed the mandelbrot in only 30 seconds on my machine, whereas
given the progress it was making, it was on pace to finish in 50
minutes. I kept both versions, so to use mine you just go to
brainfuck.h and change USING_COMMAND_STRUCTURE to 1.