added optimization and new test program
I added the optimization of making it so that if in the code there was
a chain of the same command (ex: “++++”), then instead of making four
separate commands chained next to each other, it makes one command with
a magnitude of 4. This brought the mandelbrot time down to 10 seconds.
I also added a program that I wrote a long time ago in which you input
N and then N bytes and it outputs the N bytes in sorted order.